+++ hp3parclient/client.py
@@ -2330,6 +2330,110 @@ class HP3ParClient(object):
         return self.modifyVolumeSet(set_name, action=self.SET_MEM_REMOVE,
                                     setmembers=[name])
 
+    def createSnapshotOfVolumeSet(self, name, copyOfName, optional=None):
+        """Create a snapshot of an existing Volume Set.
+
+        :param name: Name of the Snapshot. The vvname pattern is described in
+                     "VV Name Patterns" in the HP 3PAR Command Line Interface
+                     Reference, which is available at the following
+                     website: http://www.hp.com/go/storage/docs
+        :type name: str
+        :param copyOfName: The volume set you want to snapshot
+        :type copyOfName: str
+        :param optional: Dictionary of optional params
+        :type optional: dict
+
+        .. code-block:: python
+
+            optional = {
+                'id': 12,                   # Specifies ID of the volume set
+                                            # set, next by default
+                'comment': "some comment",
+                'readOnly': True,           # Read Only
+                'expirationHours': 36,      # time from now to expire
+                'retentionHours': 12        # time from now to expire
+            }
+
+        :raises: :class: `~hp3parclient.exceptions.HTTPBadRequest`
+            - INVALID_INPUT_VV_PATTERN - Invalid volume pattern specified
+        :raises: :class: `~hp3parclient.exceptions.HTTPNotFound`
+            - NON_EXISTENT_SET - The set does not exist
+        :raises: :class: `~hp3parclient.exceptions.HTTPNotFound`
+            - EMPTY_SET - The set is empty
+        :raises: :class: `~hp3parclient.exceptions.HTTPServiceUnavailable`
+            - VV_LIMIT_REACHED - Maximum number of volumes reached
+        :raises: :class: `~hp3parclient.exceptions.HTTPNotFound`
+            - NON_EXISTENT_VOL - The storage volume does not exist
+        :raises: :class:`~hp3parclient.exceptions.HTTPForbidden`
+            - VV_IS_BEING_REMOVED - The volume is being removed
+        :raises: :class: `~hp3parclient.exceptions.`
+            - INV_OPERATION_VV_READONLY_TO_READONLY_SNAP - Creating a
+            read-only copy from a read-only volume is not permitted
+        :raises: :class: `~hp3parclient.exceptions.HTTPConflict`
+            - NO_SNAP_CPG - No snapshot CPG has been configured for the volume
+        :raises: :class:`~hp3parclient.exceptions.HTTPBadRequest`
+            - INV_INPUT_DUP_NAME - Invalid input (duplicate name).
+        :raises: :class: `~hp3parclient.exceptions.HTTPForbidden`
+            - INV_OPERATION_VV_SNAP_PARENT_SAME_BASE - Two parent
+            snapshots share thesame base volume
+        :raises: :class: `~hp3parclient.exceptions.HTTPConflict`
+            - INV_OPERATION_VV_ONLINE_COPY_IN_PROGRESS - Invalid
+            operation. Online copyis in progress
+        :raises: :class: `~hp3parclient.exceptions.HTTPServiceUnavailable`
+            - VV_ID_LIMIT_REACHED - Max number of volumeIDs has been reached
+        :raises: :class:`~hp3parclient.exceptions.HTTPNotFound`
+            - NON_EXISTENT_VOLUME - The volume does not exists
+        :raises: :class:`~hp3parclient.exceptions.HTTPForbidden`
+            - VV_IN_STALE_STATE - The volume is in a stale state.
+        :raises: :class:`~hp3parclient.exceptions.HTTPForbidden`
+            - VV_NOT_STARTED - Volume is not started
+        :raises: :class: `~hp3parclient.exceptions.HTTPForbidden`
+            - VV_UNAVAILABLE - The volume is not accessible
+        :raises: :class: `~hp3parclient.exceptions.HTTPServiceUnavailable`
+            - SNAPSHOT_LIMIT_REACHED - Max number of snapshots has been reached
+        :raises: :class: `~hp3parclient.exceptions.HTTPServiceUnavailable`
+            - CPG_ALLOCATION_WARNING_REACHED - The CPG has reached the
+            allocation warning
+        :raises: :class:`~hp3parclient.exceptions.HTTPConflict`
+            - INV_OPERATION_VV_VOLUME_CONV_IN_PROGRESS - Invalid operation: VV
+            conversion is in progress.
+        :raises: :class:`~hp3parclient.exceptions.HTTPForbidden`
+            - INV_OPERATION_VV_CLEANUP_IN_PROGRESS - Internal volume cleanup is
+            in progress.
+        :raises: :class:`~hp3parclient.exceptions.HTTPForbidden`
+            - INV_OPERATION_VV_PEER_VOLUME - Cannot modify a peer volume.
+        :raises: :class:`~hp3parclient.exceptions.HTTPConflict`
+            - INV_OPERATION_VV_ONLINE_COPY_IN_PROGRESS  - The volume is the
+            target of an online copy.
+        :raises: :class:`~hp3parclient.exceptions.HTTPForbidden`
+            - INV_OPERATION_VV_INTERNAL_VOLUME - Illegal op on internal vol
+        :raises: :class: `~hp3parclient.exceptions.HTTPConflict`
+            - EXISTENT_ID - An ID exists
+        :raises: :class:`~hp3parclient.exceptions.HTTPForbidden`
+            - INV_OPERATION_VV_NOT_IN_NORMAL_STATE - Volume state is not normal
+        :raises: :class:`~hp3parclient.exceptions.HTTPForbidden`
+            - VV_IN_INCONSISTENT_STATE - Internal inconsistency error in vol
+        :raises: :class:`~hp3parclient.exceptions.HTTPBadRequest`
+            - INV_INPUT_RETAIN_GT_EXPIRE - Retention time is greater than
+            expiration time.
+        :raises: :class:`~hp3parclient.exceptions.HTTPBadRequest`
+            - INV_INPUT_TIME - Invalid time specified.
+        :raises: :class: `~hp3parclient.exceptions.`
+            - INV_OPERATION_SNAPSHOT_NOT_SAME_TYPE - Some snapshots in the
+            volume set are read-only, some are read-write
+        """
+
+        parameters = {'name': name}
+        if optional:
+            parameters = self._mergeDict(parameters, optional)
+
+        info = {'action': 'createSnapshot',
+                'parameters': parameters}
+
+        response, body = self.http.post('/volumesets/%s' % copyOfName,
+                                        body=info)
+        return body
+
     # QoS Priority Optimization methods
     def setQOSRule(self, set_name, max_io=None, max_bw=None):
         """